tutorials/AgileRL/requirements.txt
agilerl>=0.1.12
pettingzoo[atari,mpe]>=1.23.1
SuperSuit>=3.9.0
torch>=2.0.1
fastrand==1.3.0
gymnasium>=0.28.1
imageio>=2.31.1
Pillow>=9.5.0
